pyproject.toml
[project]
name = "steam-update-inhibit-sleep"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncinotify>=4.2.1",
    "dbus-python>=1.4.0",
    "inotify-simple>=1.3.5",
    "setproctitle>=1.3.6",
]
